+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+
+## [0.3.9] (2019-01-28)
+
  * Fix bug with wrong file permissions. [#49]
 
 ## [0.3.8] (2019-01-27)
@@ -70,7 +73,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
  * Initial release
 
-[Unreleased]: https://github.com/contao/image/compare/0.3.8...HEAD
+[Unreleased]: https://github.com/contao/image/compare/0.3.9...HEAD
+[0.3.9]: https://github.com/contao/image/compare/0.3.8...0.3.9
 [0.3.8]: https://github.com/contao/image/compare/0.3.7...0.3.8
 [0.3.7]: https://github.com/contao/image/compare/0.3.6...0.3.7
 [0.3.6]: https://github.com/contao/image/compare/0.3.5...0.3.6